ui: encapsulate Perfetto UI start-up in a new API class
Define a new Frontend API class that encapsulates the Perfetto UI
initialization logic with sufficient factorization of the start-up
sequence to allow an embedding application to override and customize
the Perfetto UI behavior. This includes hooks to override/customize
capabilities such as:

- adjust routing to work with applications that use hashes already
  for their own incompatible routing requirements
- customization of the content security policy
- allow to control cache keys for browser restrictions on caching
- allow file CORS origins for desktop-style applications
- scope key handling to the viewer page so that it doesn't get key
  events from other UI elements in the host application
- current script is null also in some embedder apps, not just in tests,
  so allow an application to set its own assets root URL
- suppressing the dialog that prompts to open a trace already loaded
  in the trace_processor_shell, for apps that manage this a priori
- suppressing registration of error handlers. An embedding
  application often has its own error handling

Additionally, in support of incorporating the build output into another
application's packaging:

- make @types/mithril a dev dependency
- remove unused dependencies from package.json

Fixes #2266
