+++ src/trace_processor/rpc/httpd.cc
@@ -48,6 +48,7 @@ const char* kDefaultAllowedCORSOrigins[] = {
     "https://ui.perfetto.dev",
     "http://localhost:10000",
     "http://127.0.0.1:10000",
+    "file://",
 };
 
 class Httpd : public base::HttpRequestHandler {
